编译10531：修复音频块apply作用域、绘制三角形参数遮蔽、音频块宽度类型，修复并行书签颜色行bookmark作用域，登记编译基线10531

+++ app/src/main/java/io/legado/app/ui/book/read/page/provider/TextChapterLayout.kt
@@ -863,7 +863,7 @@ class TextChapterLayout(
         var rowHeight: Float
         if (n == 1 && media[0].third == "audio") {
             // 音频块：整行、固定高度
-            cellWidth = visibleWidth
+            cellWidth = visibleWidth.toFloat()
             rowHeight = audioBlockHeight
         } else if (layoutType == BookIllustration.LAYOUT_SINGLE && displayHeight > 0f &&
             media[0].third == "image"
